More refactoring [skip hazardlib]
