+++ openquake/calculators/classical_risk.py
@@ -113,7 +113,6 @@ class ClassicalRiskCalculator(base.RiskCalculator):
                 return
         rlzs = self.datastore['csm_info'].rlzs
         self.param = dict(stats=oq.risk_stats(), weights=rlzs['weight'])
-        self.R = len(rlzs)
         self.riskinputs = self.build_riskinputs('poe')
         self.A = len(self.assetcol)
         self.L = len(self.riskmodel.loss_types)
