+++ openquake/calculators/scenario_risk.py
@@ -95,8 +95,6 @@ class ScenarioRiskCalculator(base.RiskCalculator):
         if 'gmfs' in self.oqparam.inputs:
             self.pre_calculator = None
         base.RiskCalculator.pre_execute(self)
-        if 'gmfs' in self.oqparam.inputs:
-            base.save_gmfs(self)
         self.assetcol = self.datastore['assetcol']
         A = len(self.assetcol)
         E = self.oqparam.number_of_ground_motion_fields
